Refactor #2
